Organized packages, added building blocks for an updater.
Added functionality to refresh the table with F5, so that sorting can
take place (tried to make it go automatically, but listening to changes
in every property of every member of the table list proved too difficult
for me right now). Also added possibility to delete tasks from the list,
both by hitting "delete" when task is selected and by using a context
menu. For the updater, added a library to download files, wrapped the
stage in a notificationPane and added a changelog file to the project
from which the program can get version information.