Moved control of projects and contexts to TaskManager.
Task class now has no knowledge of ProjectManager or ContextManager, and
parsing of text input or file decoding is done in TaskManager, with Task
being more of a bean (I think). Also made only one instance of Project
and Context Managers, with the control of active or inactive members
being done by it.