Added active status to Projects.
Projects now have information as to how many tasks are associated to
them and whether they are active or not. Also made some tentative
changes to make archived tasks more permanent, not allowing the user to
edit them while they are archived.
Small visual change to make the note area expand with the main content
area.