Completed the funcionality of Dropbox integration and new file system.
On launch, the program will ask every user whether to import old files
or start fresh, and also ask them to choose whether to save on dropbox
(if available) or local folder. There will be less flexibility for file
location, since the user will only be able to select the root folder of
the program. Every task is now a file, and they save and load
independently.