Changed how the program deals with files modified by itself and saving.
Now holds all changes in a 3 second period before actually saving. Also,
when a task is going to save itself, it wars the manager to ignore file
changes for a certain period. This reduces the amount of IO and also
helps avoid clashes between the task in disk and memory.