Fixed DateUtils bug, improvement to TaskManager
Setting a dueDate as "tomorrow" no longer causes a bug when done on the
last day of a month. TaskManager now only acts on changes in files after
a delay, allowing the system to stabilize before reading the file. This
must be done because a WatchService fires several events when a file is
changed, and Dropbox deletes a file and creates a new one on syncing.