Made big changes to the mechanism for saving and loading tasks.
I am trying to create a solid base for managing tasks, since my previous
attempt was all over the place and had no real consistency of functions.
Now, the Task has no knowledge of files or of the TaskManager, and the
TaskManager is responsible for managing the relationships between the
tasks, as well as keeping all of them organized and up to date.