Removed files that will no be used, added some others.
Created the main Window, which will show the available views. Also
started the projects view, that will be comprised of a tree of projects
on the left, with the selected one being fully shown on the right. For
this, I started to create cell factories for tasks, and I decided to try
to create custom components "from scratch" to be exactly what I want.